+++ src/Stryker.Core/Stryker.Core/ProjectComponents/TestProjects/TestProjectsInfo.cs
@@ -30,7 +30,6 @@ namespace Stryker.Core.ProjectComponents.TestProjects
         }
 
         public static TestProjectsInfo operator +(TestProjectsInfo a, TestProjectsInfo b)
-
         {
             a.TestProjects = a.TestProjects.Union(b.TestProjects);
             return a;
